fix compile error

+++ test/test_runner/test_runner.cpp
@@ -60,7 +60,7 @@ bool TestRunner::testStatement(TestStatement* statement, Connection& conn,
     }
     // Run the non-last queries
     size_t numParsed = parsedStatements.size();
-    for (int i = 0; i < numParsed - 1; i++) {
+    for (size_t i = 0; i < numParsed - 1; i++) {
         preparedStatement =
             conn.prepareNoLock(std::move(parsedStatements[i]), statement->enumerate);
         if (!preparedStatement->isSuccess()) {
